refactoring and fix for DOT_28673_verifyCreditCardIsNotCreatingDuplicateEntries
