+++ src/test/java/com/serenity/appium/poc/tests/DOT_26940.java
@@ -19,7 +19,7 @@ public class DOT_26940 extends WineAppTest{
         appSteps.gotoHomeTab();
         appSteps.verifyChangeStoreLookupOption();
         appSteps.verifySelectStoreFromSearchResults("Mountain View", "Mountain View");
-        appSteps.verifyShopThisStore();
+        appSteps.performShopThisStoreAction();
 
         String token = "Sweet Red Wine";
         appSteps.verifyHomepageToProductSearch();
