+++ src/test/java/com/serenity/appium/poc/tests/onHold/DOT_25238.java
@@ -1,4 +1,4 @@
-package com.serenity.appium.poc.tests;
+package com.serenity.appium.poc.tests.onHold;
 
 import com.serenity.appium.poc.WineAppTest;
 import com.serenity.appium.poc.utils.FindStore;
